Partially fix my stupidity
